backend/requirements.txt
asgiref==3.2.3
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
sendotp==0.0.4
sqlparse==0.3.1
